+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2021-10-15
+### Fixed
+- iPad loading issue.
+- support WebGL development build. Fixed emcc error of missing graphics API.
+- blurry canvas on retina and HiDPI mobile browsers
+
 ## [1.2.0] - 2021-09-22
 ### Added
 - Zappar image training in editor.